Code merging and refactoring
