+++ src/shared/components/case-viewer/case-viewer.module.ts
@@ -3,6 +3,7 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatTabsModule } from '@angular/material';
 import { RouterModule } from '@angular/router';
+import { EventManagementModule } from '..';
 import { AlertModule } from '../../../components/banners/alert';
 import { TabsModule } from '../../../components/tabs';
 import { ConditionalShowModule, LabelSubstitutorModule } from '../../directives';
@@ -22,20 +23,19 @@ import { CaseHeaderModule } from '../case-header';
 import { CaseHistoryModule } from '../case-history';
 import { ErrorMessageComponent } from '../error-message/error-message.component';
 import { ErrorsModule } from '../error/errors.module';
-import { EventTriggerModule } from '../event-trigger';
 import { ComplexModule, EventLogModule, PaletteModule } from '../palette';
 import { CaseBasicAccessViewComponent } from './case-basic-access-view/case-basic-access-view.component';
 import { CaseChallengedAccessRequestComponent } from './case-challenged-access-request/case-challenged-access-request.component';
 import { CaseChallengedAccessSuccessComponent } from './case-challenged-access-success/case-challenged-access-success.component';
-import { CaseSpecificAccessRequestComponent } from './case-specific-access-request/case-specific-access-request.component';
-import { CaseReviewSpecificAccessRequestComponent } from './case-review-specific-access-request/case-review-specific-access-request.component'
 import { CaseEventTriggerComponent } from './case-event-trigger';
 import { CaseFullAccessViewComponent } from './case-full-access-view/case-full-access-view.component';
+import { CaseReviewSpecificAccessRejectComponent } from './case-review-specific-access-reject';
+import { CaseReviewSpecificAccessRequestComponent } from './case-review-specific-access-request/case-review-specific-access-request.component';
+import { CaseSpecificAccessRequestComponent } from './case-specific-access-request/case-specific-access-request.component';
 import { CaseViewComponent } from './case-view';
 import { CaseViewerComponent } from './case-viewer.component';
 import { CasePrinterComponent, PrintUrlPipe } from './printer';
 import { CaseResolver, EventTriggerResolver } from './services';
-import { CaseReviewSpecificAccessRejectComponent } from './case-review-specific-access-reject';
 
 @NgModule({
   imports: [
@@ -44,7 +44,7 @@ import { CaseReviewSpecificAccessRejectComponent } from './case-review-specific-
     ErrorsModule,
     ActivityModule,
     CaseHeaderModule,
-    EventTriggerModule,
+    EventManagementModule,
     TabsModule,
     PaletteModule,
     LabelSubstitutorModule,
